+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.7
+
+- Update: Builtin doc was updated with latest
+  redis-doc(dd4159397f115d53423c21337eedb04d3258d291).
+- Test: IRedis now was tested in both Redis 5 and Redis 6.
+
 ### 1.6.2
 
 - Bugfix: `INFO` command accepts `section` now.